Workaround: increase git http.postBuffer in Dockerfile
There are a ton of clone operations to initialize a Zephyr project, and
a lot of them were timing out when setting up the container. In
addition, the timeouts don't cause the `west update` command to fail, so
it wastes time running only to realize a clone op has failed 10 min
later.